The analyse only considers the available genres to search for a track.
The analyse of the data is considered to suggest songs to the user. A selection of the button next throws another request with the analyse that has been done before.

NOTE : We'll have to call the makeAnalyse() function each time an user likes/dislikes a music or connects/disconnects from the host.